+++ palace/models/timeoperator.cpp
@@ -211,236 +211,10 @@ public:
   }
 };
 
-class TimeDependentCurlCurlOperator : public mfem::SecondOrderTimeDependentOperator
-{
-public:
-  // MPI communicator.
-  MPI_Comm comm;
-
-  // System matrices and excitation RHS.
-  std::unique_ptr<Operator> K, M, C;
-  Vector NegJ;
-
-  // Time dependence of current pulse for excitation: -J'(t) = -g'(t) J. This function
-  // returns g'(t).
-  std::function<double(double)> &dJ_coef;
-
-  // Internal objects for solution of linear systems during time stepping.
-  double a0_, a1_;
-  std::unique_ptr<KspSolver> kspM, kspA;
-  std::unique_ptr<Operator> A, B;
-  mutable Vector RHS;
-
-  // Bindings to SpaceOperator functions to get the system matrix and preconditioner, and
-  // construct the linear solver.
-  std::function<void(double a0, double a1)> ConfigureLinearSolver;
-
-public:
-  TimeDependentCurlCurlOperator(const IoData &iodata, SpaceOperator &space_op,
-                                std::function<double(double)> &dJ_coef, double t0,
-                                mfem::TimeDependentOperator::Type type)
-    : mfem::SecondOrderTimeDependentOperator(space_op.GetNDSpace().GetTrueVSize(), t0,
-                                             type),
-      comm(space_op.GetComm()), dJ_coef(dJ_coef)
-  {
-    // Construct the system matrices defining the linear operator. PEC boundaries are
-    // handled simply by setting diagonal entries of the mass matrix for the corresponding
-    // dofs. Because the Dirichlet BC is always homogeneous, no special elimination is
-    // required on the RHS. Diagonal entries are set in M (so M is non-singular).
-    K = space_op.GetStiffnessMatrix<Operator>(Operator::DIAG_ZERO);
-    C = space_op.GetDampingMatrix<Operator>(Operator::DIAG_ZERO);
-    M = space_op.GetMassMatrix<Operator>(Operator::DIAG_ONE);
-
-    // Set up RHS vector for the current source term: -g'(t) J, where g(t) handles the time
-    // dependence.
-    space_op.GetExcitationVector(NegJ);
-    RHS.SetSize(NegJ.Size());
-    RHS.UseDevice(true);
-
-    // Set up linear solvers.
-    {
-      auto pcg = std::make_unique<CgSolver<Operator>>(comm, 0);
-      pcg->SetInitialGuess(iodata.solver.linear.initial_guess);
-      pcg->SetRelTol(iodata.solver.linear.tol);
-      pcg->SetAbsTol(std::numeric_limits<double>::epsilon());
-      pcg->SetMaxIter(iodata.solver.linear.max_it);
-      auto jac = std::make_unique<JacobiSmoother<Operator>>(comm);
-      kspM = std::make_unique<KspSolver>(std::move(pcg), std::move(jac));
-      kspM->SetOperators(*M, *M);
-    }
-    {
-      // For explicit schemes, recommended to just use cheaper preconditioners. Otherwise,
-      // use AMS or a direct solver. The system matrix is formed as a sequence of matrix
-      // vector products, and is only assembled for preconditioning.
-      ConfigureLinearSolver = [this, &iodata, &space_op](double a0, double a1)
-      {
-        // Configure the system matrix and also the matrix (matrices) from which the
-        // preconditioner will be constructed.
-        A = space_op.GetSystemMatrix(a0, a1, 1.0, K.get(), C.get(), M.get());
-        B = space_op.GetPreconditionerMatrix<Operator>(a0, a1, 1.0, 0.0);
-
-        // Configure the solver.
-        if (!kspA)
-        {
-          kspA = std::make_unique<KspSolver>(iodata, space_op.GetNDSpaces(),
-                                             &space_op.GetH1Spaces());
-        }
-        kspA->SetOperators(*A, *B);
-      };
-    }
-  }
-
-  void FormRHS(const Vector &u, const Vector &du, Vector &rhs) const
-  {
-    // Multiply: rhs = -(K u + C du) - g'(t) J.
-    K->Mult(u, rhs);
-    if (C)
-    {
-      C->AddMult(du, rhs, 1.0);
-    }
-    linalg::AXPBYPCZ(-1.0, rhs, dJ_coef(t), NegJ, 0.0, rhs);
-  }
-
-  void Mult(const Vector &u, const Vector &du, Vector &ddu) const override
-  {
-    // Solve: M ddu = -(K u + C du) - g'(t) J.
-    if (kspM->NumTotalMult() == 0)
-    {
-      // Operators have already been set in constructor.
-      ddu = 0.0;
-    }
-    FormRHS(u, du, RHS);
-    kspM->Mult(RHS, ddu);
-  }
-
-  void ImplicitSolve(const double a0, const double a1, const Vector &u, const Vector &du,
-                     Vector &k) override
-  {
-    // Solve: (a0 K + a1 C + M) k = -(K u + C du) - g'(t) J, where a0 may be 0 in the
-    // explicit case. At first iteration, construct the solver. Also don't print a newline
-    // if already done by the mass matrix solve at the first iteration.
-    if (!kspA || a0 != a0_ || a1 != a1_)
-    {
-      // Configure the linear solver, including the system matrix and also the matrix
-      // (matrices) from which the preconditioner will be constructed.
-      ConfigureLinearSolver(a0, a1);
-      a0_ = a0;
-      a1_ = a1;
-      k = 0.0;
-    }
-    Mpi::Print("\n");
-    FormRHS(u, du, RHS);
-    kspA->Mult(RHS, k);
-  }
-};
-
 }  // namespace
 
-SecondOrderTimeOperator::SecondOrderTimeOperator(const IoData &iodata,
-                                                 SpaceOperator &space_op,
-                                                 std::function<double(double)> &dJ_coef)
-{
-  // Construct discrete curl matrix for B-field time integration.
-  Curl = &space_op.GetCurlMatrix();
-
-  // Allocate space for solution vectors.
-  E.SetSize(Curl->Width());
-  dE.SetSize(Curl->Width());
-  En.SetSize(Curl->Width());
-  B.SetSize(Curl->Height());
-  E.UseDevice(true);
-  dE.UseDevice(true);
-  En.UseDevice(true);
-  B.UseDevice(true);
-
-  // Create ODE solver for 2nd-order IVP.
-  mfem::TimeDependentOperator::Type type = mfem::TimeDependentOperator::EXPLICIT;
-  switch (iodata.solver.transient.type)
-  {
-    case config::TransientSolverData::Type::GEN_ALPHA:
-    case config::TransientSolverData::Type::DEFAULT:
-      {
-        constexpr double rho_inf = 1.0;
-        ode = std::make_unique<mfem::GeneralizedAlpha2Solver>(rho_inf);
-        type = mfem::TimeDependentOperator::IMPLICIT;
-      }
-      break;
-    case config::TransientSolverData::Type::NEWMARK:
-      {
-        constexpr double beta = 0.25, gamma = 0.5;
-        ode = std::make_unique<mfem::NewmarkSolver>(beta, gamma);
-        type = mfem::TimeDependentOperator::IMPLICIT;
-      }
-      break;
-    case config::TransientSolverData::Type::CENTRAL_DIFF:
-      {
-        ode = std::make_unique<mfem::CentralDifferenceSolver>();
-        type = mfem::TimeDependentOperator::EXPLICIT;
-      }
-      break;
-  }
-
-  // Set up time-dependent operator for 2nd-order curl-curl equation for E.
-  op =
-      std::make_unique<TimeDependentCurlCurlOperator>(iodata, space_op, dJ_coef, 0.0, type);
-}
-
-const KspSolver &SecondOrderTimeOperator::GetLinearSolver() const
-{
-  const auto &curlcurl = dynamic_cast<const TimeDependentCurlCurlOperator &>(*op);
-  MFEM_VERIFY(curlcurl.kspA,
-              "No linear solver for time-depdendent operator has been constructed!\n");
-  return *curlcurl.kspA;
-}
-
-double SecondOrderTimeOperator::GetMaxTimeStep() const
-{
-  const auto &curlcurl = dynamic_cast<const TimeDependentCurlCurlOperator &>(*op);
-  MPI_Comm comm = curlcurl.comm;
-  const Operator &M = *curlcurl.M;
-  const Operator &K = *curlcurl.K;
-
-  // Solver for M⁻¹.
-  constexpr double lin_tol = 1.0e-9;
-  constexpr int max_lin_it = 10000;
-  CgSolver<Operator> pcg(comm, 0);
-  pcg.SetRelTol(lin_tol);
-  pcg.SetMaxIter(max_lin_it);
-  pcg.SetOperator(M);
-  JacobiSmoother<Operator> jac(comm);
-  jac.SetOperator(M);
-  pcg.SetPreconditioner(jac);
-
-  // Power iteration to estimate largest eigenvalue of undamped system matrix M⁻¹ K (can use
-  // Hermitian eigenvalue solver as M, K are SPD).
-  ProductOperator op(pcg, K);
-  double lam = linalg::SpectralNorm(comm, op, true);
-  MFEM_VERIFY(lam > 0.0, "Error during power iteration, λ = " << lam << "!");
-  return 2.0 / std::sqrt(lam);
-}
-
-void SecondOrderTimeOperator::Init(double &dt)
-{
-  // Always use zero initial conditions.
-  E = 0.0;
-  dE = 0.0;
-  B = 0.0;
-  ode->Init(*op);
-}
-
-void SecondOrderTimeOperator::Step(double &t, double &dt)
-{
-  // Single time step for E-field.
-  En = E;
-  ode->Step(E, dE, t, dt);
-
-  // Trapezoidal integration for B-field: dB/dt = -∇ x E.
-  En += E;
-  Curl->AddMult(En, B, -0.5 * dt);
-}
-
-FirstOrderTimeOperator::FirstOrderTimeOperator(const IoData &iodata, SpaceOperator &space_op,
-                             std::function<double(double)> &dJ_coef)
+TimeOperator::TimeOperator(const IoData &iodata, SpaceOperator &space_op,
+                           std::function<double(double)> &dJ_coef)
 {
   // Construct discrete curl matrix for B-field time integration.
   Curl = &space_op.GetCurlMatrix();
@@ -463,25 +237,32 @@ FirstOrderTimeOperator::FirstOrderTimeOperator(const IoData &iodata, SpaceOperat
   E.MakeRef(sol, size_E);
 
   // SUNDIALS adaptive time-stepping parameters.
-  adapt_dt = iodata.solver.transient.adaptive_dt;
+  adapt_dt = true;
   rel_tol = iodata.solver.transient.rel_tol;
   abs_tol = iodata.solver.transient.abs_tol;
   order = iodata.solver.transient.order;
-  // Minimum RK order is 2.
-  order = std::max(order, 2);
 
   // Create ODE solver for 1st-order IVP.
-  mfem::TimeDependentOperator::Type type = mfem::TimeDependentOperator::EXPLICIT;
+  mfem::TimeDependentOperator::Type type = mfem::TimeDependentOperator::IMPLICIT;
   switch (iodata.solver.transient.type)
   {
+    case config::TransientSolverData::Type::GEN_ALPHA:
+    case config::TransientSolverData::Type::DEFAULT:
+      {
+        constexpr double rho_inf = 1.0;
+        use_mfem_integrator = true;
+        ode = std::make_unique<mfem::GeneralizedAlphaSolver>(rho_inf);
+        op = std::make_unique<TimeDependentFirstOrderOperator>(iodata, space_op, dJ_coef, 0.0, type);
+      }
+      break;
     case config::TransientSolverData::Type::ARKODE:
+    case config::TransientSolverData::Type::RUNGE_KUTTA:
       {
 #if defined(MFEM_USE_SUNDIALS)
         // SUNDIALS ARKODE solver.
         std::unique_ptr<mfem::ARKStepSolver> arkode;
         arkode = std::make_unique<mfem::ARKStepSolver>(space_op.GetComm(),
                                                        mfem::ARKStepSolver::IMPLICIT);
-        type = mfem::TimeDependentOperator::IMPLICIT;
         // Operator for first-order ODE system.
         op = std::make_unique<TimeDependentFirstOrderOperator>(iodata, space_op, dJ_coef, 0.0, type);
         // Initialize ARKODE.
@@ -490,19 +271,23 @@ FirstOrderTimeOperator::FirstOrderTimeOperator(const IoData &iodata, SpaceOperat
         arkode->UseMFEMLinearSolver();
         // Implicit solve is linear and J is not time-dependent.
         ARKStepSetLinear(arkode->GetMem(), 0);
-        // Maximum IRK order is 5.
-        order = std::min(order, 5);
         // Relative and absolute tolerances.
         arkode->SetSStolerances(rel_tol, abs_tol);
         // Set the order of the RK scheme.
         ARKStepSetOrder(arkode->GetMem(), order);
+        if (iodata.solver.transient.type == config::TransientSolverData::Type::RUNGE_KUTTA)
+        {
+          adapt_dt = false;
+        }
         // Set the ODE solver to ARKODE.
         ode = std::move(arkode);
+#else
+        MFEM_ABORT("Solver was not built with SUNDIALS support, please choose a "
+                   "different transient solver type!");
 #endif
       }
       break;
     case config::TransientSolverData::Type::CVODE:
-    case config::TransientSolverData::Type::DEFAULT:
       {
 #if defined(MFEM_USE_SUNDIALS)
         // SUNDIALS CVODE solver.
@@ -519,19 +304,21 @@ FirstOrderTimeOperator::FirstOrderTimeOperator(const IoData &iodata, SpaceOperat
         cvode->UseMFEMLinearSolver();
         // Set the max order of the multistep scheme.
         // CV_BDF can go up to 5, but >= 3 is not unconditionally stable.
-        order = std::min(order, 5);
         cvode->SetMaxOrder(order);
         // Set the max number of steps allowed in one CVODE step() call.
         cvode->SetMaxNSteps(10000); //default 500
         // Set the ODE solver to CVODE.
         ode = std::move(cvode);
+#else
+        MFEM_ABORT("Solver was not built with SUNDIALS support, please choose a "
+                   "different transient solver type!");
 #endif
       }
       break;
   }
 }
 
-const KspSolver &FirstOrderTimeOperator::GetLinearSolver() const
+const KspSolver &TimeOperator::GetLinearSolver() const
 {
   const auto &firstOrder = dynamic_cast<const TimeDependentFirstOrderOperator &>(*op);
   if (isExplicit())
@@ -548,7 +335,7 @@ const KspSolver &FirstOrderTimeOperator::GetLinearSolver() const
   }
 }
 
-double FirstOrderTimeOperator::GetMaxTimeStep() const
+double TimeOperator::GetMaxTimeStep() const
 {
   const auto &firstOrder = dynamic_cast<const TimeDependentFirstOrderOperator &>(*op);
   MPI_Comm comm = firstOrder.comm;
@@ -574,11 +361,15 @@ double FirstOrderTimeOperator::GetMaxTimeStep() const
   return 2.0 / std::sqrt(lam);
 }
 
-void FirstOrderTimeOperator::Init(double &dt)
+void TimeOperator::Init(double &dt)
 {
   // Always use zero initial conditions.
   sol = 0.0;
   B = 0.0;
+  if (use_mfem_integrator)
+  {
+    ode->Init(*op);
+  }
 #if defined(MFEM_USE_SUNDIALS)
   if (mfem::ARKStepSolver* arkode = dynamic_cast<mfem::ARKStepSolver*>(ode.get()))
   {
@@ -591,7 +382,7 @@ void FirstOrderTimeOperator::Init(double &dt)
 #endif
 }
 
-void FirstOrderTimeOperator::Step(double &t, double &dt)
+void TimeOperator::Step(double &t, double &dt)
 {
   En = E;
   double dt_input = dt;
@@ -604,7 +395,7 @@ void FirstOrderTimeOperator::Step(double &t, double &dt)
   Curl->AddMult(En, B, -0.5 * dt);
 }
 
-void FirstOrderTimeOperator::PrintStats()
+void TimeOperator::PrintStats()
 {
 #if defined(MFEM_USE_SUNDIALS)
   if (mfem::ARKStepSolver* arkode = dynamic_cast<mfem::ARKStepSolver*>(ode.get()))
@@ -656,7 +447,6 @@ void FirstOrderTimeOperator::PrintStats()
     Mpi::Print(" Number of error test failures: {:d}\n", netfails);
     Mpi::Print("\n");
   }
-
 #endif
 }